[FIX] move methods which have nothing to do in purchase.order.line but are related to purchase.requisition (../generic-addons/ rev 29.1.5)
